Make BytecodeBuilder into a more extensible class
In preparation for creating a JitBuilder extension for BytecodeBuilder,
to facilitate the client API creation, the BytecodeBuilder class needs
to become more like a fully extensible with an independent header.
